Optimized and reworded the algorithm
The new algorithm is more efficient and much easier for the user to read and to understand. This is now a useful function in order to calculate the factorization of prime numbers, with some errors that will show up if the number introduced by the user is not an integer higher than 1.